Move InfiniteLoopDetector to snapshot.rs
